# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AudiusSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "provider": {
            "const": "audius",
            "title": "Provider",
            "type": "string"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/AudiusTrack"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "provider",
          "tracks",
          "limit",
          "offset"
        ],
        "title": "AudiusSearchResponse",
        "type": "object"
      },
...
      "AudiusTrack": {
        "additionalProperties": true,
        "properties": {
          "artist": {
            "default": "",
            "title": "Artist",
            "type": "string"
          },
          "artwork_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artwork Url"
          },
          "can_import": {
            "title": "Can Import",
            "type": "boolean"
          },
          "duration_seconds": {
            "default": 0,
            "title": "Duration Seconds",
            "type": "number"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "import_reason": {
            "title": "Import Reason",
            "type": "string"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "permalink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permalink"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "can_import",
          "import_reason"
        ],
        "title": "AudiusTrack",
        "type": "object"
      },
...
      "DirectUploadRequest": {
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          }
        },
        "required": [
          "filename"
        ],
        "title": "DirectUploadRequest",
        "type": "object"
      },
      "DirectUploadResponse": {
        "additionalProperties": true,
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "max_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Bytes"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "object": {
            "$ref": "#/components/schemas/ObjectReference"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "method",
          "url",
          "object"
        ],
        "title": "DirectUploadResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObjectReference": {
        "additionalProperties": true,
        "properties": {
          "bucket": {
            "title": "Bucket",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "bucket",
          "key"
        ],
        "title": "ObjectReference",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Asynchronous music source-separation control plane.",
    "title": "StemSplitter API",
    "version": "development"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sources/audius/search": {
      "get": {
        "operationId": "searchAudius",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": "20",
              "title": "Limit",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": "0",
              "title": "Offset",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudiusSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Audius",
        "tags": [
          "sources"
        ]
      }
    },
    "/sources/audius/tracks/{track_id}": {
      "get": {
        "operationId": "getAudiusTrack",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audius Track",
        "tags": [
          "sources"
        ]
      }
    },
    "/uploads": {
      "post": {
        "operationId": "createDirectUpload",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Direct Upload",
        "tags": [
          "uploads"
        ]
      }
    }
  }
}